fix: construct project structure
